Add support for Client State Indications
New methods:

- client.markActive()
- client.markInactive()

These methods server as a toggle indicating to the server that it is ok
to perform traffic optimizations (such as buffering presence or other
filtering) because the client is "inactive" and such traffic would not
actually be used.

Primarily intended for mobile use cases in order to preserve battery
life by keeping the radio off as much as possible.