Don't actually send a stream error if keepalives fail.
Sending a stream error triggers a full session close, which if that
actually does make it to the server (say because the keepalive failed
only because of a very slight lag in the connection) it would prevent
the session from being resumable with stream management.

The new behaviour is to fake the stream error event so that the
application has an indication that the connection-timeout has occurred,
and kills the underlying connection without ending the session itself.