Must evaluate 'resume' in SM 'enabled' response
Consider this exchange between a stanza.io client, and a server. It shows the client enabling stream management, as described in [section 3 of XEP-0198](https://xmpp.org/extensions/xep-0198.html#enable).

raw:outgoing: `<enable xmlns="urn:xmpp:sm:3" resume="1"/>`
raw:incoming: `<enabled xmlns="urn:xmpp:sm:3"/>`

Here, the client indicates that wants to be allowed to resume a previous setting, by including a boolean 'resume' attribute.

When the server allows for Ack-ing functionality, but does not support then Resumption functionality, it is allowed to respond with an `<enabled>` element that does not include the `resume` attribute (or possibly with a falsy attribute value).

Stanza.io should, in this case, not mark the stream to be resumable, but, prior to this commit, it does.

This commit adds unit tests for this scenario and modifies the processing of the `<enabled>` response. It now explicitly evaluates the `resume` attribute of the `<enabled>` response.